Simple instruction-level coverage
Coverage calculation assumes both branches of all conditional jumps
may be taken, so in that sense it is conservative.

However, the jump tables generated by "switch" statements and accessed
via the "MOV @A, DPTR" instruction are beyond the scope of this analysis
(and thus ignored, as a conservative approach would assume any memory location
could be reached).
These will be supported by reading the debug information output by the compiler.