More efficient scheduler implementation
This scheduler now only handles the currently active blocks in the system.
When a block becomes active, it reports this to the scheduler which then
puts it in the active list. Conversely, when a block becomes idle it is
removed from this list. This means that fewer blocks are considered when
calculating the next event.